DEV: rename package, fix #11
- change project name in pyproject.toml
- change project name in README.md
  - also fix link to github repo in README.md
- rename Python package `koumura` -> `birdsongrec`
- rename test_koumura.py -> test_birdsongrec.py
