+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "koumura"
+name = "birdsong-recognition-dataset"
 description = "Functions for working with this data repository: https://figshare.com/articles/BirdsongRecognition/3470165"
 version = "0.2.1.post1"
 authors = [
@@ -38,3 +38,6 @@ Source = "https://github.com/NickleDave/koumura"
 exclude = [
     "tests/test_data"
 ]
+
+[tool.flit.module]
+name = "birdsongrec"
